requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
alibabacloud-credentials==0.2.0
alibabacloud-dingtalk==1.2.2
alibabacloud-endpoint-util==0.0.3
alibabacloud-openapi-util==0.1.5
alibabacloud-tea==0.2.8
alibabacloud-tea-openapi==0.2.8
alibabacloud-tea-util==0.3.5
async-timeout==4.0.1
asynctest==0.13.0
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==36.0.0
frozenlist==1.2.0
idna==3.3
multidict==5.2.0
pycparser==2.21
PyYAML==6.0
requests==2.26.0
typing_extensions==4.0.0
urllib3==1.26.7
xlrd==2.0.1
yarl==1.7.2